Extract test broker
